// app/mods/implementations/flood/app/scripts/core/wave.d.ts
import { Container, Sprite } from "pixi.js";
import WavesManager from "./wavesManager.ts";
import Maze from "../../../../../../../scripts/mazes/maze.ts";
import MovableEntity from "../../../../../../../scripts/characters/movableEntity.ts";
declare class Wave extends Sprite {
    speedY: number;
    startTime: number;
    started: boolean;
    decreasing: boolean;
    duration: number;
    maze: Maze;
    wavesManager: WavesManager;
    container: Container;
    elements: Sprite[];
    queuedList: any[];
    private surface;
    private surfacePhase;
    private lastOscillation;
    constructor(wavesManager: WavesManager, maze: Maze, width: number, height: number);
    queueElement(type: string, element: Exclude<Sprite, MovableEntity>): void;
    queuedElementsBy(type: string): Sprite[];
    protected dequeueElement(element: any): boolean;
    addElement(element: Sprite): void;
    removeElement(element: Sprite): void;
    clearElements(): void;
    getElementsBy(name?: string): Sprite[];
    private generateBubbles;
    private getGeneratedBubbles;
    increase(elapsedMs: number): void;
    decrease(elapsedMs: number): void;
    updatePosition(): void;
    get isDescreasing(): boolean;
    get isAtMax(): boolean;
    containsEntity(entity: any): boolean;
    cancel(): void;
    show(): void;
    draw(): void;
}
export default Wave;
//# sourceMappingURL=wave.d.ts.map